moving stories-service/ into story-list/.
