+++ antrashare/src/app/feed/story-list/stories-service/stories.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
 
-import { News } from '../../shared/models/News';
+import { News } from '../../../shared/models/News';
 
 @Injectable({
   providedIn: 'root'
